'Stairs Sweep' ver 0.04
:+1: Tuning level
:+1: Animation of dying enemies and star curtain
:+1: Sound effects
:+1: Gorgeous title logo image
:rocket: Partial screen drawing
:bug: Fix some tiny bugs

+++ stairssweep/common.cpp
@@ -20,7 +20,7 @@ MyArduboyV  arduboy;
 RECORD_T    record;
 int8_t      padX, padY, padRepeatCount;
 uint8_t     startLevel;
-uint16_t    lastScore;
+uint32_t    lastScore;
 bool        isInvalid, isRecordDirty;
 
 /*  Local Functions  */
